Moved degree-tile conversion to utils
